requirements.txt
streamlit>=1.30.0
langgraph>=0.0.20
google-genai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
# Multilingual Support
googletrans==4.0.0rc1
# Voice Interaction (Easy Install - No PyAudio!)
gTTS==2.5.0